pyproject.toml
[project]
name = "unsloth_finetuning"
version = "0.1.0"
description = "Finetuning of Gemma-2 2B for structured output"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT License" }
dependencies = [
    "Flask>=3.0.3",
    "gunicorn>=23.0.0",
    "llama-cpp-python>=0.2.88",
    "huggingface-hub>=0.24.6",
    "httpx>=0.27.0",
    "pydantic>=2.8.2",
]

[project.optional-dependencies]
test = ["pytest>=8.3.2"]

[project.urls]
respository = "https://github.com/bastienpo/unsloth_finetuning"

[tool.ruff]
# Same as black
line-length = 88

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.uv]
dev-dependencies = ["ruff>=0.6.1"]
